matcher/pattern/file/serialized: move deserialization related codes to deser.rs
